pyproject.toml
[tool.poetry]
name = "cloudflare-frontend"
version = "0.1.0"
description = ""
authors = ["Ryan Schuetzler <ryan@schuetzler.net>"]
readme = "README.md"
packages = [{include = "cloudflare_frontend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
requests = "^2.31.0"
cloudflare = "^2.12.4"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.2"
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
